+++ CHANGELOG.md
@@ -19,6 +19,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Fixed
 - Don't output a format if most string values have no format
 - Correctly check UUID format
+- Correctly check email format
 
 ## [0.11.0]
 ### Added